requirements.txt
joblib==1.1.0
matplotlib==3.4.3
networkx==2.6.3
numpy==1.17.4
pandas==1.3.4
python_igraph==0.9.8
scikit_learn==1.0.1
seaborn==0.11.2
stellargraph==1.2.1
tqdm==4.62.2
